Create m6_2WoSgiven_issn_download_all_info_about_articles
Given the ISSN of a journal, the above code downloads all the articles published in the journal, with API form Web of Science. For this code to work you must have an API Key. 

Information on how to get API key and more details can be found on the manual on the link below:
https://sites.google.com/view/fairness-in-science-in-eng/research-works